Finish Transaction Controller

+++ src/main/java/com/example/personal_budget/exception/TransactionNotFoundException.java
@@ -1,9 +1,7 @@
 package com.example.personal_budget.exception;
 
-import java.math.BigInteger;
-
 public class TransactionNotFoundException extends RuntimeException {
-  public TransactionNotFoundException(BigInteger id) {
-    super("Transaction with id " + id + " not found");
-  }
+    public TransactionNotFoundException(Long id) {
+        super("Transaction with id " + id + " not found");
+    }
 }
